Cache parser generation add more tests
Declaration is not very well defined now and it is causing
issues. It should probably be syntactic statements that
are then checked for validity after parsing, but some
declarations are expressions and some are not, so it causes
issues.

Alternatively, we could try to make all of them expressions
which basically turns several into let-binding type things.