First stab at pointing where type errors are
This was a non-trivial refactor since the inference code
had forgotten almost all about the positioning of the code.
By introducing an ad-hoc typeclass: HasRegion, we can keep
the code mostly the same (hurrah for parameteric tags hanging
on to data types).

Simplified the error cases to remove two possible errors statically.