checkpoint with more tests
It seems the previous fix was not a fix, and maybe a step in the wrong
direction. This adds a few more tests and some notes about the current
cases.

We still are not handling universal quantifications in syntax correctly
by the time we go to do inference. They are being potentially downgraded
to unknown type variables solved by unification.

We want to have a phase where we error.

One idea could be explicitly keeping a list of all universal variables
and error if it is solved?